Implement typescript binary discovery for monorepo
In a node project, the typescript binary should live under node_modules.
However, in a monorepo project where a node project can be a subproject of another
the binary of typescript could live under the node_modules of the parent project.
For this reason, the script in this commit do a recursive search in the node_modules in
all the parent projects for typescript in all the node_modules until it finds one.

Error handling can be improved